+++ apps/admin/src/app/measurement-rules/components/measurement-rule-form.tsx
