+++ apps/admin/src/app/measurement-rules/new/page.tsx
@@ -5,7 +5,7 @@ import { useRouter } from "next/navigation";
 import { useState } from "react";
 
 import { Alert, AlertDescription } from "@/components/ui/alert";
-import { MeasurementRuleForm } from "@/features/measurement-rules/components/measurement-rule-form";
+import { MeasurementRuleForm } from "@/app/measurement-rules/components/measurement-rule-form";
 import { useToast } from "@/hooks/use-toast";
 import { measurementRules, type MeasurementRule } from "@/lib/data";
