+++ apps/admin/src/app/measurement-rules/page.tsx
@@ -1,5 +1,5 @@
 import { Button } from "@/components/ui/button";
-import { MeasurementRuleList } from "@/features/measurement-rules/components/measurement-rule-list";
+import { MeasurementRuleList } from "@/app/measurement-rules/components/measurement-rule-list";
 
 export default function MeasurementRulesPage() {
   return <MeasurementRuleList />;
